Add "exact match"
It was somewhat annoying to have to select a city in the list, when all matches was pretty much the same.
E.g. "Göteborg" showing "Gothenburg, SE", "Göteborgs Stad, SE" and "Göteborg, SE".

Therefore I added a split check to show the first matching value instead of a list when searching for e.g. "Göteborg, SE".